requirements.txt
amqp==5.0.6
appdirs==1.4.4
argh==0.26.2
asgiref==3.4.1
astroid==2.6.6
billiard==3.6.4.0
black==21.7b0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-environ==0.4.5
djangorestframework==3.12.4
flake8==3.9.2
flower==1.0.0
humanize==3.11.0
idna==3.2
isort==5.9.3
kombu==5.1.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.9.6
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.8.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
tornado==6.1
urllib3==1.26.6
vine==5.0.0
watchdog==2.1.3
wcwidth==0.2.5
wrapt==1.12.1
